+++ bevy_asset_loader/examples/atlas_from_grid.rs
@@ -7,14 +7,17 @@ use bevy_asset_loader::prelude::*;
 fn main() {
     App::new()
         .add_state::<MyStates>()
+        .add_plugins(DefaultPlugins)
         .add_loading_state(
             LoadingState::new(MyStates::AssetLoading).continue_to_state(MyStates::Next),
         )
         .add_collection_to_loading_state::<_, MyAssets>(MyStates::AssetLoading)
         .insert_resource(Msaa::Off)
-        .add_plugins(DefaultPlugins)
-        .add_system(draw_atlas.in_schedule(OnEnter(MyStates::Next)))
-        .add_system(animate_sprite_system.run_if(in_state(MyStates::Next)))
+        .add_systems(OnEnter(MyStates::Next), draw_atlas)
+        .add_systems(
+            Update,
+            animate_sprite_system.run_if(in_state(MyStates::Next)),
+        )
         .run();
 }